Implement inbound control and query API
Also:

hotSampleWithVolumeUpdates (v1.1.0) adds de1_time. de1_time and
scale_time are preferred over arrival_time as, in a future version,
these will be estimates that remove some of the jitter relative to
packet-arrival time.

To be able to keep cached values of DE1 variables current, a read-back
is requested on each write.

NoneSet and NONE_SET added to some enum.IntFlag to provide clearer
representations

Although is_read_once and is_stable have been roughed in,
optimizations using them have not been done

Disabled reads of CUUID.ReadFromMMR as it returns the request itself
(which is not easily distinguashable from the data read. These two
interpret their Length field differently, making it difficult to
determine if 5 is an unexpected value or if it was just that 6 words
were requested to be read.

Scaling on MMR0x80LowAddr.TANK_WATER_THRESHOLD was corrected.