Use Singleton class for DE1, FlowSequencer, and ScaleProcessor
This class derived from that of Guido van Rossum available at
https://www.python.org/download/releases/2.2.3/descrintro/

Changed "init" to "_singleton_init" for clarity in subclasses

As noted there and in the code, classes implementing Singleton
should *not* define __init__() but rather _singleton_init()

This naming change should also help avoid otential, future confusion
between "set up this singleton once" and "reset all the state to be
ready for reconnection".

The optional address argument was removed from DE1 and Scale
as the singleton might be created elsewhere first, leading to
silent failure to initialize the address. Although Scale is not
a sigleton, keep the initialization of the two consistent.

DE1.address and Scale.address properties now accept either a string
with the address or a BLEDevice such as returned from BleakScanner,
at least as fully as bleak supports it. At this time, bleak seems
unable to accept the BleakClientBlueZDBus object returned under Linux.
The address setter extracts the address string as a special case.