Improve atexit shutdown of Bluetooth connections, BleakClientWrapped
The BleuZ stack will hold on to connections after the app exits.
This can result in needing to manually disconnect the devices,
either with bluetoothctl, or by restarting bluetooth.servie.

As certain devices may have problems without an explicit disconnect,
provide disconnect for all platforms.

The previous scheme for atexit handlers was unified under
BleakClientWrapped, which is now used instead of BleakClient,
from which it inherits.

BleakClientWrapped also provides a .name property, which replaces
the code in commit a5ef96a. Should the bleak distribution provide
BleakClient.name in the future, this definition should be removed.

BleakClientWrapped.willful_disconnect is provided so that the
on-disconnect callback can differentiate between an intentional
and non-intentional disconnect. Although the bleak documentation
suggests that the callback is only for unintentional disconnects,
this does not appear to be the case in practice.

The logging of the specific Scale class was simplified to only
the class name, such as AtomaxSkaleII, rather than repr()