Determine and use connectivity requirements for GET/PATCH requests
By determining the requirements for a specific GET or PATCH request,
they can be processed in cases when, for example, a Scale is not
connected, or requesting that connectivity be established.

  NB: Although the DE1 and Scale can be connected,
      they are not reinitialized at this time.

API CHANGES
===========

MAPPING VERSION 2.0.0
---------------------

Adds `IsAt.internal_type` to help validate the string values for
`DE1ModeEnum` and `ConnectivityEnum`. JSON producers and consumers
should sill expect and provide `IsAt.v_type`