Provide profile upload through API
Upload of profile (JSON "v2" format) available with PUT at de1/profile

  curl -D - -X PUT --data @examples/jmk_eb5.json \
  http://localhost:1234/de1/profile

At this time, the upload is synchronous from the perspective of the caller.
The upload of the example profile (5 frames) takes about 1.7 seconds.
Additional frames typically add about 100 ms each.

If/when the behavior of this PUT is changed to be asynchronous,
both a notification of completion and an ability to cancel
will be provided. Please check CHANGELOG.md for such changes.

API CHANGES
===========

MAPPING VERSION 2.1.0
---------------------

Enables `de1/profile` for PUT