Support non-GHC machines for mode commands
Machines without a GHC rely on the API to initiate flow,
Espresso, HotWaterFLush, Steam, and HotWater.

Extended DE1.mode_setter() to allow these mode-change commands if the
GHC is known not to be installed. If MMR0x80LowAddr.GHC_INFO has not
yet been retrieved, is is await-ed prior to processing the request.

NB: This has not been tested on a machine wihout a GHC

Added HTTP response code to be able to distinguish between "bad timing"
and "this is not going to be possible" (due to GHC, firmware, ...)

Resolved pickling issue with custom exception by defining __reduce__()
Pickle of custom exceptions discussed at
    https://stackoverflow.com/questions/16244923/
    how-to-make-a-custom-exception-class-with-multiple-init-args-pickleable

API CHANGES
===========

RESOURCE VERSION
----------------

1.2.0 -- add flow-control modes for non-GHC use

RESPONSE CODES
--------------

409 -- When the current state of the device does not permit the action
       DE1APIUnsupportedStateTransitionError

418 -- When the device is incapable of or blocked from taking the action
       DE1APIUnsupportedFeatureError