Move "controller" to its own process
The controller is now running in its own process in preparation for
process supervision and centralized logging.

Much of what was in `try_de1.py` is now in `controller.py`

The structure of this file is still under refactoring.
The intent is that the "standard" initialization and setup
will be a single call. This will likely be part of managing
connection, reconnection, and change of device.