Refactor IPC between controller and MQTT to use a pipe
Use of loop.add_reader() eliminates the need for running blocking code
with loop.run_in_executor() and the complexity of gracefully interrupting
that `while True` loop.

There is only a single writer and the objects are effectively
written atomically. The additional overhead of a queue seems
unneeded.

At this time there is no lock on the writer end of the pipe.

An intermediate asyncio.Queue was considered to help mitigate the impact
of the MQTT process not picking up messages quickly enough. This was
not implemented at this time.