Provide centralized file logging with SIGHUP for reopen
Each process now logs to a multiprocessing.Queue and a thread of the
main process copies the entries to the log file.

By default, the log file is `/tmp/log/pyDE1/combined.log`
and is appended to.

SIGHUP will close the file and reopen, for use with OS-supplied or other
log-rotation tools.

Several, "internal" calls have had signature changes to accommodate the
changes in IPC.

While functional, this is a waypoint commit prior to refactoring
for removal of repetition and clarity of organization.