Long-running processes, tasks, and futures are supervised
These processes now log and restart themselves if they abort

The DE1 and Scale both "self-initialize" on connect.

Much of `try_de1.py` has been moved into `controller.py`

This is a waypoint commit to be able to capture operational code
prior to refactoring for maintainability and clarity.

Known Issues:

* A "permanent" failure results in continuous respawning

* Exceptions on a non-supervised task or callback are "swallowed"
  by the default Python asyncio exception handler. They are logged,
  but do not terminate execution.

* The main process can return a non-zero code even when it appears that
  shutdown was intentional and graceful

* The API for enabling and disabling auto-tare and stop-at can only do
   so within the limits of the FlowSequencer's list of applicable states.
   See further `autotare_states`, `stop_at_*_states`, `last_drops_states`