Move the ugly_bits out, reintroduce try_de1.py as an example
There are still a few tasks associated with setup that aren't handled
by the new run(). These "ugly bits" are imported by module name,
hopefully allowing them to be overridden with a path-based import of

    async def manual_setup(disconnect_set: set)

as

    from ugly_bits import manual_setup

The path loaded is logged

    INFO [Controller] Controller: Scheduled task for manual_setup()
        from /full/path/to/ugly_bits.py

try_de1.py is quite simple now. pyDE1/run.py can also be run as main

NB: Do not forget to use `if __name__ == '__main__':`

    from pyDE1.run import run

    if __name__ == '__main__':
        run()