Provide a timeout for the HTTP server waiting for controller
Previously, the HTTP server would block forever if the controller
did not respond. RESPONSE_TIMEOUT from pyDE1.config.http now is used.

The response pipe is now cleared when the process starts, just before
the server is enabled. In the case of a restart, an unserviced response
might be "left over". Otherwise this left-over response would be
wrongly returned to the next request, as would all subsequent.

A blocking connection.poll() is used as the servicing of requests
is synchronous, sequential, and running in a dedicated process.
It is hard to justify more complexity at this time.