Robustly track ready/not_ready, allow query through inbound API
Although the ConnectivityEvent messages indicated readiness of the DE1
and the Scale, if a client was not connected at the time that the device
became ready, it previously had no way of confirming "ready" prior to sending
commands. Although this is usually only a fraction of second for a scale and
a couple of seconds for the DE1, this meant that a client could not use "ready"
as a signal that its operations could continue.

Try to make sure that at least one not_ready is sent when the device
starts and begins to disconnect or is disconnected. At least one ...

API CHANGE
==========

RESOURCE VERSION 2.0.0

Added `ConnectivityEnum.READY = 'ready'`

The connectivity queries to `Resource.DE1_CONNECTIVITY` and
`Resource.SCALE-CONNECTIVITY` will now return `ConnectivityEnum.READY`
 in addition to `.NOT_CONNECTED` and `.CONNECTED`.

Previous code that assumed that `.CONNECTED` was the terminal state
should be modified to recognize `.READY`.