Update examples/find_first_and_load.py to show tracking "ready"
With "ready" available from the connectivity query, it is now possible
to check if the device is already connected, connect if not, and then
wait until the device is ready for further commands.

Implementation in this example is done with the paho and request modules
and "classic" threading is used.