Modify profile storage and generation for id, fingerprint, and source
Retain the source in as-received form for capture into database

Define a profile id as the SHA1 of the source

Define a profile fingerprint as the SHA of the frame set sent to the DE1

To calculate the "fingerprint" of the profile's instruction set, for frame-based profiles:

* Convert the profile into the `ShotDescHeader`, `ShotFrame`, `ShotExtFrame`,
  and `ShotTail` series to be sent.
* Concatenate the over-the-wire, binary payload of those in the following order:
	* `ShotDescHeader`
	* `ShotFrame` in ascending order
	* `ShotExtFrame` in ascending order
	* `ShotTail`
* Take the SHA1 of the result

Although the `ShotExtFrame` is "optional" if it is "empty" and
should not change the operation of the system, the implementation here
will be to use what is sent, neither adding or removing frames.

Internal Changes:

`Profile.from_json()` now expects a string or bytes-like object,
rather than a dict. This change is to ease capture of the
profile "source" for use with history logging.

`ProfileByFrames.from_json()` no longer rounds the floats to maintain
the integrity of the original source. They will still be rounded
at the time that they are encoded into binary payloads.

Deprecated (Internal)

`Profile.from_json_file()` as it is no longer needed with the API
able to upload profiles. If needed within the code base, read
the file, and pass to `Profile.from_json()` to ensure that
the profile source and signatures are properly updated.