Provide Visualizer-compatible "shot file" data
Extract a sequence and its set of records from the database,
based on sequence ID. Create a string that is suitable for upload
to Visualizer or for use with John Weiss' shot-plotting tools.

This is intended for backwards compatibility with existing tools,
not as an interchange format for the future.