Provide stand-alone app that automatically uploads to Visualizer
Monitoring MQTT packets, this example app detects than an "Espresso"
has completed, creates a "shot file" through access to the SQLite3
database, and uploads it to Visualizer.

The results are reported over MQTT, including the Visualizer URL
on success.

It runs independently from the pyDE1 processes and will listen for
shot after shot, even if those processes are restarted.