Correct reporting not-connected conditions to HTTP API
When the DE1 or scale is required to be connected to answer the query
but it is not, a 409 (Conflict) response should be returned.

In two cases, this was not happening. This resolves a timeout when
requesting from the DE1 when not connected, as well as a null response
involving a scale when one is not configured.

DE1NotConnectedError exceptions are now logged as INFO
as they generally shouldn't represent a coding or operational error

Fixes #4