Improve API query behavior when DE1 or Scale is not ready
In general, if the DE1 or the Scale isn't connected and ready,
you can't do much with them. You do need to be able to
set address as well as connect.

Allow: DE1_ID, DE1_CONNECTIVITY, SCALE_ID, and SCALE_CONNECTIVITY

Add both DE1.is_ready() and Scale.is_ready() and change the check
from "connected" to "ready"

Collect together logic for more consistent and robust implementation.

Properly reflect that the DE1 needs to be ready to read
from PackedAttr or MMR.