Allow [dis]connect while [dis]connected
Don't force the GUI to check if a device is already disconnected
before trying to disconnect it.

Still some issues with the assert when DE1 is not connected,
probably due to an intentional disconnect of the DE1 readying it
for connection to a potentially "new" DE1 and nulling its address.
"Why else would you intentionally disconnect?"

(Scale seems OK for this)

  File "/home/ble-remote/devel/pyDE1/src/pyDE1/de1/de1.py",
      line 1224, in connectivity_setter
    await self.connect()
  File "/home/ble-remote/devel/pyDE1/src/pyDE1/de1/de1.py",
       line 395, in connect
    assert self._bleak_client is not None
AssertionError

2021-07-31 08:03:23,285 INFO [Controller] DE1: Connecting to DE1 at None
2021-07-31 08:03:23,291 ERROR [Controller] Dispatcher:
    Exception in processing HTTPMethod.PATCH
        Resource.DE1_CONNECTIVITY AssertionError()