Database Schema 2 -- capture id, name to connectivity_change
NB: Backup your database before applying this

    See, for example, the SQLite .backup command

Adds two columns to connectivity_change to capture the Bluetooth ID
and name of the device, now being sent in the MQTT update.

Capture data into those two columns, as available.

upgrade.001.002.sql is not an automated trigger.

This trigger does not confirm that the database is at schema 1.