Exception, signal, and shutdown utilities
Asyncio has an unfortunate behavior of "swallowing" exceptions
and not killing the process (by default). As a result, without "help",
an exception won't terminate and allow a service supervisor to restart.

Similar challenges exist with threads (either explicit, or when using
run_in_executor), as well as that threads can't be "killed" in a robust
way from a running program (if at all).

This commit provides utilities that can assist in the capture of
exceptions, as well as module-level flags (effectively a singleton)
to enable graceful shutdown.

Utilities are also provided to manage shutdown of running processes,
including the default executor, asyncgens, and the running asyncio loop.

threading.Event: shutdown_manager.shutdown_underway

can be checked by threads to determine when they should shutdown and
exit. This is likely to migrate to multiprocessing.Event shortly.

THe shutdown will wait for up to CLEANUP_WAIT seconds (default, 5) for

threading.Event: shutdown_manager.cleanup_complete

before shutting down. This allows processes and threads to wait on
shutdown_manager.shutdown_underway, then do what is needed for a
graceful shutdown, before setting shutdown_manager.cleanup_complete.
Coalescing these actions into one routine is suggested, as it is a
threading.Event as opposed to some type of "counting" lock. As with
shutdown_underway, this is likely to migrate to a
multiprocessing.Event in the near future.