pyde1-visualizer runs as a systemd service
Using the shutdown_manager from a previous commit, the auto-upload
to Visualizer has been converted to a service. A systemd unit file
for the service is provided.

Configuration is now using TOML (don't worry, it looks like a classic
".ini" file).

Once the unit file is linked and enabled, it will start as system boot
and be restarted in the event of an unexpected failure.