Use TOML config files outside of the pyDE1 distribution
To enable easier installation, configuration, and maintenance of pyDE1
services, the previous Python-based "parameter" files have been migrated
to TOML (".ini-like") files. Default locations are in /usr/local/etc/pyde1/
and will be able to be overridden with a -c <config_file_path> argument
on starting the app, either from the command line or as a service.

This will allow local configuration to not be overridden if the user
updates the pyDE1 module. New configuration items will have default values
even if not set in the local configuration file(s).

TOML looks a lot like .ini or some sectioned config files and is relatively
forgiving of formatting. One quirk is that there is no representation for
"None" or "null" available. As a result, the parameters that allow None
in Python are generally set to None as a default. Not specifying an override
in the config file will then allow use of None at run time.

In most cases, the default values should be sufficient. The code should run
with the TOML config files missing. Notable exceptions are credentials,
such as those for Visualizer uploads.