Database will self-initialize, check database schema version
Assuming that the configured directory for the database exists
and is writable for the user, the database will be created
and initialized with the current schema.

If there is a schema mismatch, the app will, at this time, exit.
Schema updates will be addressed in the future, when needed.

A utility for timestamped backup of the database is provided
and run if the database file existed already, but did not have
a user_version set. It uses the sqlite3 .backup command run in
a subprocess. When complete, the backup is compressed by running
a subprocess in the default executor.

New config parameters in the [database] section include:

  BACKUP_TIMEOUT = 60  # seconds
  BACKUP_COMPRESSION_EXECUTABLE = 'xz'