Refactor shutdown process and adapt processes
NB: SIGHUP is no longer used for log rotation and is a "kill" signal

Introduce a generalized process-shutdown framework and utilize across
the set of processes. The shutdown_manager module also includes an
exception handler for the loop, as well as the shutdown_if_exception()
callback that can be used on asyncio tasks.

The general shutdown process now is that (imported as 'sm')

  * sm.shutdown_underway gets set
  * SIGTERM is deemed a "graceful" request
  * Wait up to sm.CLEANUP_WAIT seconds for sm.cleanup_complete
    (This allows the process to shut down, the indicate readiness)
  * Shuts down the default_executor, cancels tasks, stops the loop
  * and exits with:
    * os.EX_OK if "graceful"
    * -signal value if from another signal
    * os.EX_SOFTWARE if called directly (without a signal)

Logging has been moved to using a QueueListener which simplifies
shutdown. Formatting is done by the listener's handlers, so "remote"
loggers should not need to format with their QueueHandler.

This replaces signal_handlers.py, which has been removed.