Bluetooth connection tracking for Linux
On Linux, Bluetooth devices are not "cleaned up" by the OS when a process exits.
Although both the bleak library and this code attempts to ensure that they
are not left in a connected state by the OS's BlueZ stack, there are conditions
when this may still happen, such as SIGKILL or low-level faults. On relaunch,
these devices generally will not be advertising (as they are connected) and
can't be connected.

These functions manage files in config.bluetooth.ID_FILE_DIRECTORY, suffixed by
config.bluetooth.ID_FILE_SUFFIX, that may be used by a supervisory script
to disconnect any that remain connected.

After extracting the device ID from the file and properly sanitizing it,

    bluetoothctl disconnect D9:B2:48:aa:bb:cc

can run by a member of the bluetooth group to disconnect the device.