SupervisedProcess: Wrap execution in a try/except and shutdown on exception
If an uncaught exception occurs in a multiprocessing process,
the process may not exit on its own.

Wrap the execution of the target in a try/except. If there is an exception,
call sm.shutdown(None, loop) with either the running loop, or if not running,
by obtaining a loop and run_until_complete()