Enable use of username/password auth for MQTT client
With auth and ACLs, access to publishing on the MQTT server can be restricted.
As these are, I believe, transmitted in the clear in the CONNECT message,
appropriate care should be taken that they are not "snooped". At the present,
configuration of TLS for MQTT has not been enabled through the config files.
Until TLS is available, use of localhost is at least a modest improvement.