replay: Log MQTT connection and publish results
At least log errors when trying to connect or publish.

This can occur if the MQTT broker isn't up and running,
or the credentials are wrong.

TODO: Determine the various failure modes and how to address them.

The broker not running is addressed marginally for the services
by starting the pyde1 and pyde1-visualizer services "After" the
mosquitto service in their unit files. This does not guarantee
that the mosquitto service is up and running.

A later commit addresses connectivity and authentication errors
for the pyDE1 outbound MQTT API