config: Convert config-parsing code from TOML to YAML
TOML doesn't support null/None which will become troublesome quickly.
It already requires setting defaults of None internally and requiring
external config to override them, or some non-standard, "special" value
that is used to denote None. There are also problems with setting a key
of, for example, 'service.subservice', the name of a potential logger.

Move to YAML instead, which is still human-readable and easily editable.