FlowSequencer: Handle steam-time management without uncaught exception
When notifying a disabled stop-at-time state, no longer try to obtain
the value that would be used if it were enabled.

Use DE1APINotManagedHereException to indicate that the check for wait time
is not needed for steam, as it is handled directly by the DE1 firmware.

Also log the state when starting a sequence.