pyde1.service: Run disconnect-btid.sh to "reset" Bluez, if needed
Bluez will "hold onto" a device even after the process exits. While pyDE1
tries to close connections, an ungraceful exit may leave them connected.
If this happens, a subsequent run won't be able to use the devices.

If the .btid files are left behind, try to disconnect with bluetoothctl
and remove file if sucessful. Also try before, if they are still there.