Update c_api with new information, implement flush control API
With experimental firmware 1283, HotWaterRinse ("flush") can be controlled
through MMRs as to its target duration, temperature, and flow.

Update c_api with new information and reflect in feature_flags

Read the firmware version early on connecting to the DE1 and use it
when deciding how to read the MMR bank.

Refactor ModeControl into FlowSequencer.py, then move out of subdirectory.
As use of the firmware-provided stop-at-time for flush will require DE1 access,
this avoids the messiness of circular imports by including in FlowSequencer.py

Flush control implemented. The lower limit is 3 seconds in FW 1283.