Provide asynchronous firmware upload through API
Enable upload of a firmware file through the API as a PUT to DE1_FIRMWARE

File is loaded asynchronously to the DE1 which should be able to continue
to be used as the file is loaded. The firmware file should be sent "as-is",
simplest as application/octet-stream (multipart forms are not decoded).

Load can be canceled with a PUT to DE1_FIRMWARE_CANCEL

Progress updates are available through MQTT as FirmwareUpdate events.
These events include a FirmwareUploadState as well as current bytes uploaded
and total bytes to upload. These updates are sufficient to drive a <progress>
indicator or similar. At this time, they are sent at start, every 1 kB,
and at finish (completed, failed, or canceled).

MAPPING_VERSION  3.3.0
RESOURCE_VERSION 3.3.0