Rewrite logging and config loading
The logging system and loading of config from YAML have been rewritten
to be able to support configuration through the config files.

Notable end-user changes include:

* The config section for logging has been restructured

* All "names" are now case-sensitive

* Logging levels need to be either the name, such as 'INFO' or an int
  * Setting to 'NOTSET' is how to remove a previously-set level (not None)

* Third-party libraries get their logger from the root logger
  * They are rendered in the logs and managed as 'root.third.party'

* pyDE1 components get their logger from pyDE1.getLogger('some.name')
  * They are rendered in their logs and managed as 'some.name'

* Early logging, before the log file has been opened, is preserved
  and sent to the log file once the logging system is configured

The basic structure of the new logging system allows loggers from third-party
libraries to inherit from the root logger and for pyDE1 loggers to inherit
from pyDE1.getLogger().

For multiprocess applications, the logging is done with a QueueHandler
and a QueueListener, with delivery handlers attached to the QueueListener.

For single-process applications, the delivery handlers are attached
directly to the root logger.

There are two handlers provided:

* LOGFILE - output to the log file
* STDERR  - output to stderr

The handlers manage the formatting and filtering by level, once a log record
is in the logging chain.

The two "root" loggers are set to the smallest of the levels of these
two handlers. These set the "effective level" of child loggers that are not
explicitly set. It is the effective level of the logger that has its method
called to create a log record that determines if a record gets in to the chain.
It is the handler that determines if it is rendered to the output.

Here is the current logging section in pyDE1.conf:

logging:
    LOG_DIRECTORY: /var/log/pyde1/
    # NB: The log file name is matched against [a-zA-Z0-9._-]
    LOG_FILENAME: pyde1.log
    formatters:
        STYLE: '%'  # Both need to be the same style
        LOGFILE: >-
            %(asctime)s %(levelname)s [%(processName)s] %(name)s: %(message)s
        STDERR: >-
                        %(levelname)s [%(processName)s] %(name)s: %(message)s
    handlers:
        LOGFILE:    INFO        # The log file
        STDERR:     WARNING     # Captured by systemd
    LOGGERS:
        # What gets allowed out is further filtered by the handlers
        # logging.getLogger() and pyDE1.getLogger() are set to
        # the lowest level of the logfile and stderr handlers
        # when those handlers are set, leaving overrides for
        # overly chatty libraries if at DEBUG
        # NB: If there are no entries here, comment-out LOGGERS entirely
        root.aiosqlite:     INFO
        root.asyncio:       INFO
        root.bleak:         INFO

The LOGGERS section, if present, is used to configure specific loggers' levels.

NB: If there are no entries for LOGGERS:, a YAML syntax error will occur.
    The easiest resolution is to comment out # LOGGERS: entirely.

The three shown set the level for their packages' logging, which inherit from
their respective package-specific loggers. These packages log "too much"
information at the DEBUG level. By setting the level for the logger from which
they inherit to INFO, the records don't get into the chain, so aren't rendered
by the handlers, even if those handlers are at the DEBUG level.