MAPPING 4.0.0 - IsAt uses TO enum rather than the classes themselves
Originally, classes were used for the DE1, FlowSequencer, ... to indicate
the target of the setter/getter. To try to keep from requiring the majority
of the code being imported into processes, stubs were used when the "real"
object's module wasn't available. While working on rationalizing imports,
it became clear that this approach was not robust to changes in import order.

To resolve this the TO class is now used as a token and resolved
to an instance of the intended class in implementation.py

Removed: stubs.py