MQTT "Will" implemented, reporting unexpected disconnects
The MQTT Last Will and Testament feature is used to provide notification
of an unexpected disconnect.

pyDE1/status/mqtt/logging Died
pyDE1/status/mqtt/notification Died

NB: The intent of this feature is to reliably indicate loss of MQTT.
    That it *may* indicate if it is present is a bonus.

class MQTTStatusText (enum.Enum):
    on_connection = 'Here'
    on_graceful_disconnect = 'Gone'
    on_will = 'Died'

'Gone' should be published if the disconnect() method is called.
The call to the class-specific disconnect() will wait_for_publish()
before calling the super-class' disconnect() method.

NB: There may be other paths that disconnect that this does not intercept.
    Should this be the case, 'Here' would persist.