Enable read of Calibration, log, and access on DE1
The Calibration CUUID is supported and is read on connection to a DE1.
Retrieval is asynchronous, not delaying "ready" of the DE1.

Retrieval needs to be serial, as queued-up requests may get "lost"
without a corresponding notification from the DE1.

Calibration values are saved into DE1.cal_local and DE1.cal_factory
in normalized form, as a ratio for flow and pressure, as an offset
for temperature (degrees C).

The WriteKey for the DE1 would be required to be able to either set values
or return all to factory settings. Neither write nor reset have been tested.