Correct double-counting of scale delays
The scale delay was being removed in AtomaxSkaleII
_scale_time_from_latest_arrival() as well as in
Estimator.estimate(), along with half the nominal period.
As a result, the scale data was ~0.4 seconds early
relative to the DE1 data.

Removed the compensation in Estimator.estimate()
and confirmed that AtomaxSkaleII._sensor_lag = 0.38
includes both the estimated measurement lag of 0.33
as well as half the nominal period (0.1 / 2 = 0.05)