Fix DE1 instantiation not getting config-file values
Reported that the DE1 wasn't picking up a change to 50 Hz in the config.

Although most instances of config in the child threads appeared to have
the proper values, when the DE1 singleton was instantiated in
the controller thread, it appeared to be the default values,
not those from the load of the config.

Resolved by explicitly importing and then updating pyDE1.config.config
within the "run" method of each of the processes.

Closes #20