Lengthen timeout for setting profile by ID
The timeout for uploading a profile was config.http.PROFILE_TIMEOUT
(4.5 sec) when setting a profile by ID, the timeout was not increased.
As a result, profiles with many frames would time out and fail to load.

This commit increases the timeout to config.http.PROFILE_TIMEOUT
for both 'set_profile_by_id' and 'upload_json_v2_profile',
resolving the issue.