profiles: Factor persist of last profile out of db.insert.profile()
In preparation for being able to upload a profile to the database
when the DE1 is not connected, move persistence of last profile
into the profile-upload task itself.

This should also properly persist the last profile when selecting by ID.
The database notion of last profile is used when a recently connected DE1
reports an espresso sequence prior to having had a profile uploaded
during the current connection. In this case, the sequence indicates that
the profile in use is a guess.

Relates to #23