de1: scale: Quiet all connection attempt/fail logging when "not logging"
For the DE1 and Scale, when trying to reconnect after an unexpected
disconnection, the first `config.bluetooth.RECONNECT_RETRY_COUNT`
attempts should be logged, then continue without logging the attempts
and failures. As devices, especially battery-powered scales, may be off
for most of the day, this is intended to prevent meaningless log spam.

This patch fixes the omission of test for if these log messages
should be shown for the DE1 and Scale.

Closes #24