MAPPING VERSION 4.1.1: Changes / DEPRECATED: "first_if_found", "id", "scan"
The semantics of starting a scan or connecting to the first are changed.

To start a scan, the parameter has been changed to prefer a positive number
for the timeout, or null (to accept the default). Use of a bool here
has been DEPRECATED. The preferred forms include:

    {'begin': null}
    {'begin': 5}
    {'begin': 5.0}

To start a scan and select the first-found device of the desired type,
set the id to 'scan'. Use of the 'first_if_found' key has been DEPRECATED:
The preferred forms include:

    {'id': 'scan'}
    {'id': 'aa:bb:cc:dd:ee:ff'}
    {'id': null}

---

DEPRECATED forms include:

    {'begin': true}
    {'begin': false}
    {'first_if_found': true}
    {'first_if_found': false}