Add SEQUENCE_WATCHDOG_TIMEOUT and logic around leaving sequence "early"
Previously, it was possible that the end of a sequence was not detected.
As a result, the recorder could run until the start of the next sequence,
which might be the next day or later. One condition appears to have been
stopping a flow sequence before the flow begins. For example, deciding to
hit stop during "warm up" and before the flow starts. This previously
could result in database bloat.

To resolve this, state transitions into a non-flow state while a sequence
is running, but not yet transitioned to flow, now terminates the sequence.

A watchdog timer is also added with a duration set by
    config.de1.SEQUENCE_WATCHDOG_TIMEOUT = 270 # seconds
if the watchdog timeer fires during a sequence, the sequence is terminated.
No commands are sent to the DE1 to alter its state.

270 seconds should allow for 30 seconds of pre-flow warm-up
and stabilization and then 240 seconds of flow time.
This should be sufficient for most expresso as well as "Filter 2.1".